Fixed #33: moved user-related preferences and logic to a dedicated app
