+++ dynamic_preferences/__init__.py
@@ -1,4 +1,2 @@
-# from .dynamic_preferences_registry import user_preferences_registry, global_preferences_registry
-
 __version__ = "0.8.4"
 default_app_config = 'dynamic_preferences.apps.DynamicPreferencesConfig'
