+++ dynamic_preferences/admin.py
@@ -3,8 +3,8 @@ from django import forms
 
 from .settings import preferences_settings
 from .registries import global_preferences_registry
-from .models import GlobalPreferenceModel, UserPreferenceModel
-from .forms import GlobalSinglePreferenceForm, UserSinglePreferenceForm, SinglePerInstancePreferenceForm
+from .models import GlobalPreferenceModel
+from .forms import GlobalSinglePreferenceForm, SinglePerInstancePreferenceForm
 
 
 class DynamicPreferenceAdmin(admin.ModelAdmin):
@@ -38,16 +38,3 @@ class PerInstancePreferenceAdmin(DynamicPreferenceAdmin):
     raw_id_fields = ('instance',)
     form = SinglePerInstancePreferenceForm
     list_select_related = True
-
-class UserPreferenceAdmin(PerInstancePreferenceAdmin):
-    search_fields = ['instance__username'] + DynamicPreferenceAdmin.search_fields
-    form = UserSinglePreferenceForm
-    changelist_form = UserSinglePreferenceForm
-
-    def get_queryset(self, request, *args, **kwargs):
-        # Instanciate default prefs
-        getattr(request.user, preferences_settings.MANAGER_ATTRIBUTE).all()
-        return super(UserPreferenceAdmin, self).get_queryset(request, *args, **kwargs)
-
-if preferences_settings.ENABLE_USER_PREFERENCES:
-    admin.site.register(UserPreferenceModel, UserPreferenceAdmin)
