+++ dynamic_preferences/registries.py
@@ -201,13 +201,9 @@ class PreferenceRegistry(persisting_theory.Registry):
 class PerInstancePreferenceRegistry(PreferenceRegistry):
     pass
 
+
 class GlobalPreferenceRegistry(PreferenceRegistry):
     def populate(self, **kwargs):
         return self.models(**kwargs)
 
-class UserPreferenceRegistry(PerInstancePreferenceRegistry):
-    pass
-
-
-user_preferences_registry = UserPreferenceRegistry()
 global_preferences_registry = GlobalPreferenceRegistry()
