+++ dynamic_preferences/urls.py
@@ -1,6 +1,5 @@
 from django.conf.urls import include, url
 from django.contrib.admin.views.decorators import staff_member_required
-from django.contrib.auth.decorators import login_required
 from . import views
 from .registries import global_preferences_registry
 from .forms import GlobalPreferenceForm
@@ -19,10 +18,5 @@ urlpatterns = [
             form_class=GlobalPreferenceForm)),
         name="dynamic_preferences.global.section"),
 
-    url(r'^user/$',
-        login_required(views.UserPreferenceFormView.as_view()),
-        name="dynamic_preferences.user"),
-    url(r'^user/(?P<section>[\w\ ]+)$',
-        login_required(views.UserPreferenceFormView.as_view()),
-        name="dynamic_preferences.user.section"),
+    url(r'^user/', include('dynamic_preferences.users.urls')),
 ]
