+++ dynamic_preferences/users/registries.py
@@ -1,0 +1,8 @@
+from ..registries import PerInstancePreferenceRegistry
+
+
+class UserPreferenceRegistry(PerInstancePreferenceRegistry):
+    pass
+
+
+user_preferences_registry = UserPreferenceRegistry()
