+++ dynamic_preferences/views.py
@@ -1,6 +1,5 @@
 from django.views.generic import TemplateView, FormView
-from .forms import preference_form_builder, user_preference_form_builder
-from .registries import user_preferences_registry
+from .forms import preference_form_builder
 
 
 """Todo : remove these views and use only context processors"""
@@ -42,14 +41,3 @@ class PreferenceFormView(FormView):
 
         form.update_preferences()
         return super(PreferenceFormView, self).form_valid(form)
-
-class UserPreferenceFormView(PreferenceFormView):
-    """
-    Will pass `request.user` to form_builder
-    """
-    registry = user_preferences_registry
-
-    def get_form_class(self, *args, **kwargs):
-        section = self.kwargs.get('section', None)
-        form_class = user_preference_form_builder(instance=self.request.user, section=section)
-        return form_class
