+++ example/example/settings.py
@@ -55,6 +55,7 @@ INSTALLED_APPS = (
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'dynamic_preferences',
+    'dynamic_preferences.users.apps.UserPreferencesConfig',
     'debug_toolbar',
     'example',
 )
