+++ tests/test_app/dynamic_preferences_registry.py
@@ -1,6 +1,6 @@
 from dynamic_preferences.types import *
-from dynamic_preferences.registries import (
-    user_preferences_registry, global_preferences_registry)
+from dynamic_preferences.registries import global_preferences_registry
+from dynamic_preferences.users.registries import user_preferences_registry
 from .models import BlogEntry
